+++ CHANGELOG.md
@@ -1,2 +1,5 @@
+# Unreleased (v0.1.1)
+* Add `print-completions` command.
+
 # v0.1.0
 * Add `vcs`, `extract`, `join` commands.